feat: TS-parity overhaul — agentteam/compact/memory/skills/@mention
- core: /rewind picker reaches MessageSelector parity; vercel-ai seam
  (services/inference is the only entry, Message family relocated to
  coco-messages); SessionRuntime unifies TUI/SDK runner wiring; engine.rs
  split 4511 → 1550 LoC across builder/session/compaction/finalize_turn/
  prompt; workspace clippy-zero baseline with Stop-hook + pre-commit gate.

- memory: port auto-memory + auto-dream subsystems (extract / dream /
  session services under MemoryRuntime, gated by Feature::AutoMemory and
  routed through ModelRole::Memory); turn-end fan-out from QueryEngine,
  /dream + /summary sentinels, SideQueryAdapter for the recall ranker.

- agentteam: extract 21 swarm modules into coco-coordinator (root layer);
  spawn-time identity resolution (definition.model_role + per-spawn
  model precedence + ThinkingLevel + SubagentRuntimeSnapshot for cache
  parity); SpawnMode::{Fresh,Fork,Resume}; background AgentTool drives
  the engine, registers with TaskManager, persists per-agent JSONL
  transcript, supports tail-read + opt-in periodic AgentSummary; model-
  driven resume via SendMessageTool auto-dispatch; TUI permission
  bridge wires Ask decisions to overlay; mailbox/tmux/handoff TS-parity
  fixes; per-agent MEMORY.md / hooks / skills / MCP inheritance;
  /agents slash command; Feature::AgentTeams promoted to Experimental.

- compact: Round 10 normalize pipeline (5 ordered passes mirroring TS,
  smoosh-before-sanitize, request-id-keyed assistant merge, image
  stripping in tool-result arrays); post-compact plan-mode + async-agent
  attachments; RecompactionInfo plumbed through CompactRunOptions;
  SessionStart hooks run after LLM compact; Tool Result Budget Phase 0
  config stub + Level-1 + Level-2 tool-result spill to disk.

- skills/commands/plugins: Round 11 — bundled inventory + per-process
  nonce extraction with O_NOFOLLOW|O_EXCL, TS-faithful prompt token
  renderer, gated_by/skill_root surface; CommandResult::{Compact,Prompt,
  OpenDialog,Skip} taxonomy + DialogSpec, verbatim TS prompt files,
  TS-ordered build_command_registry seam; PluginId/PluginScope, full
  TS-port DFS dependency resolver, NFKD homograph fold for impersonation
  detection, BuiltinPluginDefinition registry, ZIP/mcpb loader with
  sha256 cache + traversal guard, 24-variant PluginError.

- context/tui/query: @-mention TS-parity overhaul — parser fragment +
  agent-suffix + symbol fixes, MentionType::McpResource variant, TUI
  matcher swapped to coco-file-search (nucleo + git ls-files / ripgrep)
  with SharedFileIndex pre-warm + git_index_watcher refresh,
  accept_suggestion auto-quoting / dir-trailing-slash / agent-prefix
  handling, mention-resolver paths resolved against cwd before reaching
  the AlreadyReadFile silent-reminder + nested-CLAUDE.md lookup keys.

- conventions: add Config & Feature Gates principles to root CLAUDE.md
  (consume RuntimeConfig only; Feature is a coarse capability gate;
  three resolution layers fold once; gate at Tool::is_enabled,
  subagents never widen parent).

+++ coco-rs/app/tui/src/widgets/notification.rs
@@ -124,7 +124,7 @@ fn kitty_id() -> u32 {
     use std::time::UNIX_EPOCH;
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
-        .map(|d| (d.subsec_nanos() % 10_000) as u32)
+        .map(|d| d.subsec_nanos() % 10_000)
         .unwrap_or(0)
 }
